requirements.txt
accelerate==0.26.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.2
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6  
coloredlogs==15.0.1
configobj==5.0.8
configparser==7.0.0
contourpy==1.1.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.16.1
deepdiff==7.0.1
dill==0.3.7
distro==1.9.0
easyocr==1.7.1
emoji==2.12.1
etelemetry==0.3.1
filelock==3.13.1
filetype==1.2.0
fitz==0.0.1.dev2
flake8-polyfill==1.0.2
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
imageio==2.34.2
importlib_resources==6.4.0
iopath==0.1.10
isodate==0.6.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.16
langchain-community==0.0.17
langchain-core==0.2.38
langchain-ollama==0.1.3
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.111
layoutparser==0.3.4
lazy_loader==0.4
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.0
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
numpy==1.24.4
ollama==0.3.2
onnx==1.16.1
onnxruntime==1.18.1
openai==1.43.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
ordered-set==4.1.0
orjson==3.10.7
packaging==23.2
panda==0.3.1
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.2.0
pinecone==4.0.0
pinecone-client==4.1.2
pinecone-plugin-interface==0.0.7
pkgutil_resolve_name==1.3.10
portalocker==2.10.1
protobuf==5.27.2
prov==2.0.1
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-bidi==0.5.2
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
rapidfuzz==3.9.5
rdflib==6.3.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.2
scikit-image==0.21.0
scipy==1.10.1
semantic-text-splitter==0.14.0
shapely==2.0.5
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.36.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tifffile==2023.7.10
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tornado==6.4.1
tqdm==4.66.1
traits==6.3.2
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
unstructured==0.11.8
unstructured-client==0.25.0
unstructured-inference==0.7.36
urllib3==2.2.0
watchdog==4.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.19.2
